requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.2
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework-stubs==3.16.8
greenlet==3.3.2
idna==3.11
iniconfig==2.3.0
packaging==26.0
pillow==12.1.1
playwright==1.58.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pyee==13.0.1
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.12.0
requests==2.32.5
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
urllib3==2.6.3
